+++ lib/state_machine/integrations/data_mapper.rb
@@ -178,6 +178,14 @@ module StateMachine
     # *not* because a matching transition was not available, no error messages
     # will be added to the state attribute.
     # 
+    # In addition, if you're using the <tt>ignite!</tt> version of the event,
+    # then the failure reason (such as the current validation errors) will be
+    # included in the exception that gets raised when the event fails.  For
+    # example, assuming there's a validation on a field called +name+ on the class:
+    # 
+    #   vehicle = Vehicle.new
+    #   vehicle.ignite!       # => StateMachine::InvalidTransition: Cannot transition state via :ignite from :parked (Reason(s): Name cannot be blank)
+    # 
     # == Scopes
     # 
     # To assist in filtering models with specific states, a series of class
@@ -325,6 +333,20 @@ module StateMachine
       # Adds a validation error to the given object
       def invalidate(object, attribute, message, values = [])
         object.errors.add(self.attribute(attribute), generate_message(message, values)) if supports_validations?
+      end
+      
+      # Describes the current validation errors on the given object.  If none
+      # are specific, then the default error is interpeted as a "halt".
+      def errors_for(object)
+        if object.errors.empty?
+          'Transition halted'
+        else
+          errors = []
+          object.errors.each_pair do |field_name, field_errors|
+            field_errors.each {|error| errors << "#{field_name} #{error}"}
+          end
+          errors * ', '
+        end
       end
       
       # Resets any errors previously added when invalidating the given object